update the server
separate the routes of the request for 2 files for now
