+++ src/apis/database.py
@@ -1,0 +1,2 @@
+from flask_pymongo import PyMongo
+mongo = PyMongo()
